+++ tests/testthat/test-dataset.R
@@ -3,7 +3,7 @@ context("Dataset object and methods")
 with(fake.HTTP, {
     test.ds <- loadDataset("test ds")
     test.ds2 <- loadDataset("ECON.sav") # assume it's OK to overwrite this below
-    today <- Sys.Date()
+    today <- "2016-02-11"
     
     test_that("Dataset can be created", {
         expect_true(is.dataset(test.ds))
@@ -13,20 +13,28 @@ with(fake.HTTP, {
         expect_identical(name(test.ds), "test ds")
         expect_identical(description(test.ds), "")
         expect_identical(id(test.ds), "511a7c49778030653aab5963")
-        expect_identical(start_date(test.ds), "2016-01-01")
-        expect_identical(end_date(test.ds), "2016-01-01")
-        expect_identical(start_date(test.ds2), NULL)
-        expect_identical(end_date(test.ds2), NULL)
+        expect_identical(startDate(test.ds), "2016-01-01")
+        expect_identical(endDate(test.ds), "2016-01-01")
+        expect_identical(startDate(test.ds2), NULL)
+        expect_identical(endDate(test.ds2), NULL)
     })
 
-    test_that("start_date makes right request", {
-        expect_error(start_date(test.ds2) <- as.character(today),
-            paste0('PATCH /api/datasets.json \\{"/api/datasets/dataset3.json":\\{"start_date":"',as.character(today),'"\\}\\}')
+    test_that("startDate is type character", {
+        expect_identical(class(startDate(test.ds)), "character")
+    })
+
+    test_that("endDate is type character", {
+        expect_identical(class(startDate(test.ds)), "character")
+    })
+    
+    test_that("startDate makes right request", {
+        expect_error(startDate(test.ds2) <- today,
+            paste0('PATCH /api/datasets.json \\{"/api/datasets/dataset3.json":\\{"startDate":"',today,'"\\}\\}')
         )
     })
-    test_that("end_date makes right request", {
-        expect_error(end_date(test.ds2) <- today,
-            paste0('PATCH /api/datasets.json \\{"/api/datasets/dataset3.json":\\{"end_date":"',today,'"\\}\\}')
+    test_that("endDate makes right request", {
+        expect_error(endDate(test.ds2) <- today,
+            paste0('PATCH /api/datasets.json \\{"/api/datasets/dataset3.json":\\{"endDate":"',today,'"\\}\\}')
         )
     })